Title: 최소 통과 시간, Time: 125ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 250914/최소 통과 시간/minimum-transit-time.java
@@ -12,7 +12,7 @@ public class Main {
         // Please write your code here.
 
         long left = 1;
-        long right = maxHour*m;
+        long right = maxHour*(long)m;
 
         while(left<right){
             long mid = (left+right)/2;
